Split search routines into new module
